requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# WebSocket
websockets==12.0

# Data Models
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment
python-dotenv==1.0.0

# Async support
aiofiles==23.2.1

# AI/ML dependencies
google-generativeai==0.8.5

# Optional: Add other AI/ML dependencies here
# openai==1.10.0
# anthropic==0.8.1
# langchain==0.1.0
# Pillow==10.2.0  # For image generation


